refactor(domain): move VerticalSliceRepository into vertical-slice subfolder
