+++ src/core/domain/vertical-slice/VerticalSliceRepository.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceCollection } from './vertical-slice/VerticalSliceCollection';
+import { type VerticalSliceCollection } from './VerticalSliceCollection';
 
 /** Port interface for vertical slice persistence — used by command handlers to load and save the collection. */
 export interface VerticalSliceRepository {
