+++ src/core/usecases/commands/DeleteBoundedContext/DeleteBoundedContextCommandHandler.ts
@@ -1,5 +1,5 @@
 import { type BoundedContextRepository } from '../../../domain/bounded-context/BoundedContextRepository';
-import { type VerticalSliceRepository } from '../../../domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../../domain/vertical-slice/VerticalSliceRepository';
 import { type DeleteBoundedContextCommand } from './DeleteBoundedContextCommand';
 
 export class DeleteBoundedContextCommandHandler {
