+++ src/core/usecases/commands/DeleteSlice/DeleteSliceCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceRepository } from '../../../domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../../domain/vertical-slice/VerticalSliceRepository';
 import { DeleteSliceCommand } from './DeleteSliceCommand';
 
 export class DeleteSliceCommandHandler {
