+++ src/core/usecases/commands/ExtendSliceRight/ExtendSliceRightCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceRepository } from '../../../domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../../domain/vertical-slice/VerticalSliceRepository';
 import { ExtendSliceRightCommand } from './ExtendSliceRightCommand';
 
 export class ExtendSliceRightCommandHandler {
