+++ src/core/usecases/commands/RemoveScenarioFromSlice/RemoveScenarioFromSliceCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceRepository } from '../../../domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../../domain/vertical-slice/VerticalSliceRepository';
 import { RemoveScenarioFromSliceCommand } from './RemoveScenarioFromSliceCommand';
 
 export class RemoveScenarioFromSliceCommandHandler {
