+++ src/core/usecases/commands/RenameSlice/RenameSliceCommandHandler.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceRepository } from '../../../domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../../domain/vertical-slice/VerticalSliceRepository';
 import { RenameSliceCommand } from './RenameSliceCommand';
 
 export class RenameSliceCommandHandler {
