+++ tests/helpers/InMemoryVerticalSliceRepository.ts
@@ -1,4 +1,4 @@
-import { type VerticalSliceRepository } from '../../src/core/domain/VerticalSliceRepository';
+import { type VerticalSliceRepository } from '../../src/core/domain/vertical-slice/VerticalSliceRepository';
 import { VerticalSliceCollection } from '../../src/core/domain/vertical-slice/VerticalSliceCollection';
 
 /** In-memory implementation of VerticalSliceRepository for testing command handlers. */
